Wow. I got a lot of stuff done in the past few hours. There's a lot of polish in this app now.
Google Drive support is the biggest feature. You can enable it (but it's not automatically enabled) when you start the app. Create a file which will then be synced across all your devices. CRUD on channels in app immediately affect the file elsewhere.

If you want instead, you can pick an already existing file to sync from.

Editing files is really simple, it's just JSON. Saving it on your PC, as long as you update the "modified" attribute, will sync it automatically to the TV within the hour. The app chooses the most recent version of the file to sync.

The channel overlay has gotten significant improvements. Now it displays the channel number, name, and thumbnail (along with a progress circle).

What's really beautiful is Palette taking the color from the thumbnail and using that to color all the widgets in this overlay. It makes each channel unique in a consistent way, very material.

The main activity has gotten a handful of new options for advanced file management.

SettingsManager has GDrive support built-in now, so that's really neat. It's not perfect by any means, but it can serve as a really useful way to keep settings synced, a neat feature.