I went through the issues list and added a few more things. Unfortunately, I did not get to plugin support. It hasn't on the top of the list and frankly it seems a little intimidating.
I improved the setup screen (close #20) giving a more transparent view of what's happening and not immediately closing because that didn't make a lot of sense.

I'm going pretty hard with resyncing, doing it several times per operation now so that Live Channels is always in sync (close #17)

File op directions were reversed (close #16), and I reworded them to be more helpful

Users should now be able to set custom splashscreens for a channel by adding the `splashscreen` attribute to a channel, navigating to an image URL that will appear when the channel is first loading. It uses Picasso, so I'm not entirely sure about animation support, but there is no spinner. (Close #14)

Genres have been introduced in the backend (#6)

I modified the OOFTV channel to the latest suggestions (close #1)

Finally, the users can now see the version number (close #19)