Made changes to leanback card presenters
There is an improved CardPresenter which uses
the native JsonChannel and is more colorful.
There is a new presenter just for the various
options in the app.
The details fragment also has visual improvements.
Now the main fragment will refresh when it receives
events, not after a set time.
Also created the SuggestedChannels class.

Close #164
Close #145
Close #93
Close #138