Add EWTN and Telecare to suggested channels
These are a couple of my mother's favorite channels. They have a variety
of content, including news, so I am marking them as news channels. I
consider them to be no less news channels than those from the United
States' big 3 broadcasters, which are considered to be news channels,
despite only airing the news a fraction of the time. There are other
channels in this genre that I could propose, but I do not feel like
tracking down logos and urls for them, so I am limiting my first patch
to these two.

I am not a lawyer, but I believe that these streams are completely legal
to distribute. They are both owned by non-profits that:

1. Want anyone to be able to watch their content
2. State that clearly on their website.
3. Make their content available free to watch on their websites.
4. Have streaming apps on at least 1 platform.

I am suggesting that these channels use channel numbers 135 and 137
because that is what the local cable company uses.

Also, the project's license is unclear. I had been very hesitant to
upstream a patch to a project without clarification on the license used
by the project. After some soul searching, I have decided to publish
this patch under the terms of either the Apache 2.0 or MIT licenses. If
the project lead wishes to fix the unclear licensing with a different
OSS license, I would likely be happy to follow suit. However, this patch
itself is so small that I doubt it qualifies for copyright protection,
so the license under which I choose to release it is likely a moot point
unless I start sending more patches.

The irony of continuing to say that my changes are too small to qualify
for copyright protection on my third proposed patch does not escape me.
They probably do when all 3 patches are taken together, but I do not
particularly care too much about the threshold at which copyright
protection applies. I just want to stress the importance of having clear
licensing and to urge Nick to adopt a clear license ASAP.

Signed-off-by: Richard Yao <ryao@gentoo.org>